Finished toggle button 4

+++ project/green_board/button_demo/stateMachines.h
@@ -2,7 +2,7 @@
 #define stateMachine_included
 
 // These should only be changed in switches.c.
-extern const int periods[], periods_inv[], time_set[];
+extern const short periods[], periods_inv[], time_set[];
 
 /* States will only be changed in switches.c*/
 extern short state_button_1, state_button_2, state_button_3, state_button_4;
